Pitch_mode  and frequency spectrum changes
Pitchmode functionality added
Exciter Knob addded with functionilty
Volume Slider added
Key-MIDI binding
Frequency Spectrum Changes

+++ Source/PresetManager.cpp
@@ -190,7 +190,7 @@ bool PresetManager::loadPreset(const juce::File& presetFile,
         for (const auto& prop : paramsObj->getProperties())
         {
             juce::String paramID = prop.name.toString();
-            float normValue = (float)prop.value;
+            float normValue = static_cast<float>(static_cast<double>(prop.value));
 
             if (auto* param = apvts.getParameter(paramID))
             {
